Extend Decidim Initiatives Admin Controller
